finished game layout, added go back buttons in the instructions and the game screen
